requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
serpapi>=2.4.0